changed date constant to function getCurrentTimeAndDate
